stb#314  Refactor test utils file location
