+++ catalog/helpers.py
@@ -5,18 +5,20 @@ from django.urls import reverse
 from catalog.models import AbstractCategory, TagQuerySet
 
 
-def get_category_path(
-    category: AbstractCategory,
-    route_name='category',
+def reverse_catalog_url(
+    url: str,
+    route_kwargs: dict,
     tags: TagQuerySet=None,
     sorting: int=None,
     query_string: dict=None,
-):
+) -> str:
     query_string = f'?{urlencode(query_string)}' if query_string else ''
-    route_kwargs = {
-        'category_id': category.id,
-        **({'tags': tags.as_url()} if tags else {}),
-        **({'sorting': sorting} if sorting is not None else {}),
-    }
+    if tags:
+        # PyCharm's option:
+        # noinspection PyTypeChecker
+        tags_slug = tags.as_url()
+        route_kwargs['tags'] = tags_slug
+    if sorting is not None:
+        route_kwargs['sorting'] = sorting
 
-    return f'{reverse(route_name, kwargs=route_kwargs)}{query_string}'
+    return f'{reverse(url, kwargs=route_kwargs)}{query_string}'
